requirements.txt
django_redis==4.11.0
asgiref==3.2.7
Django==3.0.5
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.1
timeout-decorator==0.4.1
cryptography
psycopg2-binary
